# Optimize lax.associative_scan, reimplement cumsum, etc. on top of associative_scan.