--- a/docs/jax.lax.rst
+++ b/docs/jax.lax.rst
@@ -57,6 +57,10 @@ Operators
     conv_transpose
     cos
     cosh
+    cummax
+    cummin
+    cumprod
+    cumsum
     digamma
     div
     dot